dz5, commit with realization of services and some adjustments before testing

+++ src/main/resources/db/changelog/changes/v2022062201__create_person_table.sql
@@ -2,7 +2,7 @@ CREATE SCHEMA ulab_edu;
 
 create table ulab_edu.person
 (
-    id        integer     not null,
+    id        bigint      not null,
     full_name varchar(50) not null,
     title     varchar(50) not null,
     age       integer     not null,
