+++ src/main/resources/db/changelog/changes/v2022062301__create_book_table.sql
@@ -1,6 +1,6 @@
 create table ulab_edu.book
 (
-    id         integer     not null,
+    id         bigint      not null,
     person_id  integer     not null,
     title      varchar(50) not null,
     author     varchar(50) not null,
